Add files via upload
This project is a beginner-level Machine Learning model that predicts forest cover types based on various geographical and environmental features. The goal is to classify the forest into specific types using supervised learning techniques. It serves as a hands-on exercise to get started with data preprocessing, visualization, model building, and evaluation.